Make project node startup times configurable
Add two new variables which can be used to configure the number of
attempts to check erlang project node availability (at start up) as
well as the interval between each check.

This is a version of c963f194851152bc4edbbdc7b9e4458a7a0e544c, but for
project nodes instead of the main edts node.

* edts-api-num-project-node-start-retries (default 5)
  The number of retries to wait for a project node to start before giving up

* edts-api-project-node-start-retry-interval (default 0.5)
  Time between each project node availability check at start up

This can be useful when the project's start-command takes more than
2.5 s (5 * 0.5s) to complete (slow file system, ...).