Multiuser fix reads the port number from file
The previous naive patch would allow the user to set the port number
on both the Erlang side as well as the Elisp side. The Erlang side read
the port from a file in the priv-path. On the Elisp side however, the
user would need to customize the port-variable.

This fix makes Elisp read the port number from the same file as Erlang,
thus making it more maintainable for the user.