Ignore errors when fetching the list of nodes on startup
On boot, it takes a while until the EDTS Erlang node is actually
started and ready to accept request. The current implementation
of the `edts-api-start-server` function fails in its attempt to
perform retries in the case the Erlang node is too slow to start.
In fact, the `edts-api-get-nodes` function provokes an error and
evaluation is stopped. Instead, we may want to ignore errors, so that
we can block until the server is up and running. This avoid the
notorious error:

    open-network-stream: make client process failed: Connection
    refused, :name, 0, :buffer, #<killed buffer>, :host, 0, :service,
    4587, :nowait, nil, :tls-parameters, nil

That many EDTS users seem to experience nowadays.